# openapi.yaml
openapi: 3.1.0
info:
  title: Trustify
  description: Software Supply-Chain Security API
  license:
    name: Apache License, Version 2.0
    identifier: Apache-2.0
  version: 0.5.0-beta.2
paths:
...
  /api/v3/group/sbom:
    get:
      tags:
      - sbomGroup
      summary: List SBOM groups
      operationId: listSbomGroups
      parameters:
      - name: totals
        in: query
        description: return the total number of children
        required: false
        schema:
          type: boolean
      - name: parents
        in: query
        description: return the parent chain
        required: false
        schema:
          type: string
          enum:
          - skip
          - id
          - resolve
      - name: offset
        in: query
        description: |-
          The first item to return, skipping all that come before it.

          NOTE: The order of items is defined by the API being called.
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: limit
        in: query
        description: |-
          The maximum number of entries to return.

          Zero means: return no items (the total count is still computed if requested).
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: total
        in: query
        description: Whether to compute and return the total count of matching items.
        required: false
        schema:
          type: boolean
      - name: q
        in: query
        description: |
          EBNF grammar for the _q_ parameter:
          ```text
              q = ( values | filter ) { '&' q }
              values = value { '|', values }
              filter = field, operator, values
              operator = "=" | "!=" | "~" | "!~" | ">=" | ">" | "<=" | "<"
              value = (* any text but escape special characters with '\' *)
              field = (* must match an entity attribute name *)
          ```
          Any values in a _q_ will result in a case-insensitive "full
          text search", effectively producing an OR clause of LIKE
          clauses for every string-ish field in the resource being
          queried.

          Examples:
          - `foo` - any field containing 'foo'
          - `foo|bar` - any field containing either 'foo' OR 'bar'
          - `foo&bar` - some field contains 'foo' AND some field contains 'bar'

          A _filter_ may also be used to constrain the results. The
          filter's field name must correspond to one of the resource's
          attributes. If it doesn't, an error will be returned
          containing a list of the valid fields for that resource.

          An ASCII value of `NUL`, percent-encoded as `%00`, may be used
          to find resources on which a particular field isn't set. For
          example, `name=%00` and `name!=%00` yield the WHERE clauses,
          'NAME IS NULL' and 'NAME IS NOT NULL', respectively.

          Examples:
          - `name=foo` - entity's _name_ matches 'foo' exactly
          - `name~foo` - entity's _name_ contains 'foo', case-insensitive
          - `name~foo|bar` - entity's _name_ contains either 'foo' OR 'bar', case-insensitive
          - `name=` - entity's _name_ is the empty string, ''
          - `name=%00` - entity's _name_ isn't set
          - `published>3 days ago` - date values can be "human time"

          Multiple full text searches and/or filters should be
          '&'-delimited -- they are logically AND'd together.

          - `red hat|fedora&labels:type=cve|osv&published>last wednesday 17:00`

          Fields corresponding to JSON objects in the database may use a
          ':' to delimit the column name and the object key,
          e.g. `purl:qualifiers:type=pom`

          Any operator or special character, e.g. '|', '&', within a
          value should be escaped by prefixing it with a backslash.
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: |
          EBNF grammar for the _sort_ parameter:
          ```text
              sort = field [ ':', order ] { ',' sort }
              order = ( "asc" | "desc" )
              field = (* must match the name of entity's attributes *)
          ```
          The optional _order_ should be one of "asc" or "desc". If
          omitted, the order defaults to "asc".

          Each _field_ name must correspond to one of the columns of the
          table holding the entities being queried. Those corresponding
          to JSON objects in the database may use a ':' to delimit the
          column name and the object key,
          e.g. `purl:qualifiers:type:desc`
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Executed the SBOM group query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupListResult'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
    post:
      tags:
      - sbomGroup
      summary: Create a new SBOM group
      operationId: createSbomGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupRequest'
        required: true
      responses:
        '201':
          description: Created the requested group
          headers:
            location:
              schema:
                type: string
              description: The relative URL to the created resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResponse'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '409':
          description: The name of the group is not unique within the parent
  /api/v3/group/sbom-assignment:
    put:
      tags:
      - sbomGroup
      summary: Bulk update SBOM group assignments
      operationId: bulkUpdateSbomGroupAssignments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkAssignmentRequest'
        required: true
      responses:
        '204':
          description: The SBOM assignments were updated
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
  /api/v3/group/sbom-assignment/{id}:
    get:
      tags:
      - sbomGroup
      summary: Get SBOM group assignments
      operationId: readSbomGroupAssignments
      parameters:
      - name: id
        in: path
        description: The ID of the SBOM
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The SBOM was found and assignments returned
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '404':
          description: The SBOM was not found
    put:
      tags:
      - sbomGroup
      summary: Update SBOM group assignments
      operationId: updateSbomGroupAssignments
      parameters:
      - name: id
        in: path
        description: The ID of the SBOM to update
        required: true
        schema:
          type: string
      - name: if-match
        in: header
        description: The revision of the SBOM assignments
        required: false
        schema:
          type:
          - string
          - 'null'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        required: true
      responses:
        '204':
          description: The SBOM assignments were updated
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '412':
          description: The requested revision is not the current revision
  /api/v3/group/sbom/{id}:
    get:
      tags:
      - sbomGroup
      summary: Read the SBOM group information
      operationId: readSbomGroup
      parameters:
      - name: id
        in: path
        description: The ID of the group to read
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The group was found and returned
          headers:
            etag:
              schema:
                type: string
              description: Revision ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
    put:
      tags:
      - sbomGroup
      summary: Update an SBOM group
      operationId: updateSbomGroup
      parameters:
      - name: id
        in: path
        description: The ID of the group to update
        required: true
        schema:
          type: string
      - name: if-match
        in: header
        description: The revision to update
        required: false
        schema:
          type:
          - string
          - 'null'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupRequest'
        required: true
      responses:
        '204':
          description: The group was delete or did not exist
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '409':
          description: Assigning the parent would create a cycle
        '412':
          description: The requested revision is not the current revision of the group
    delete:
      tags:
      - sbomGroup
      summary: Delete an SBOM group
      operationId: deleteSbomGroup
      parameters:
      - name: id
        in: path
        description: The ID of the group to delete
        required: true
        schema:
          type: string
      - name: if-match
        in: header
        description: The revision to delete
        required: false
        schema:
          type:
          - string
          - 'null'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupRequest'
        required: true
      responses:
        '204':
          description: The group was deleted or did not exist
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '409':
          description: The group has child groups and cannot be deleted
        '412':
          description: The requested revision is not the current revision of the group
...
components:
  schemas:
...
    BulkAssignmentRequest:
      type: object
      description: Request to assign multiple SBOMs to the same set of groups.
      required:
      - sbom_ids
      - group_ids
      properties:
        group_ids:
          type: array
          items:
            type: string
          description: The group IDs to assign to each SBOM (replaces existing assignments).
        sbom_ids:
          type: array
          items:
            type: string
          description: The IDs of the SBOMs to update.
...
    CreateResponse:
      type: object
      required:
      - id
      properties:
        id:
          type: string
          description: The ID of the newly created group
...
    Group:
      type: object
      required:
      - id
      - name
      properties:
        description:
          type:
          - string
          - 'null'
          description: A user friendly description
        id:
          type: string
          description: The ID of the group
        labels:
          $ref: '#/components/schemas/Labels'
          description: Additional group labels
        name:
          type: string
          description: The name of the group, in the context of its parent
        parent:
          type:
          - string
          - 'null'
          description: The direct parent of this group
    GroupListResult:
      allOf:
      - $ref: '#/components/schemas/PaginatedResults_GroupDetails'
      - type: object
        properties:
          referenced:
            type:
            - array
            - 'null'
            items:
              $ref: '#/components/schemas/Group'
            description: |-
              Groups referenced by parent chains but not present in the primary result set.

              Only present when `parents=resolve` is requested.
      description: Result of listing SBOM groups, with optional resolved parent references.
    GroupRequest:
      type: object
      description: Mutable properties of a [`Group`].
      required:
      - name
      properties:
        description:
          type:
          - string
          - 'null'
          description: A user provided description
        labels:
          $ref: '#/components/schemas/Labels'
        name:
          type: string
          description: The name of the group.
        parent:
          type:
          - string
          - 'null'
          description: The ID of the group's parent.
...
    Labels:
      type: object
      additionalProperties:
        type: string
...
    PaginatedResults_GroupDetails:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            allOf:
            - $ref: '#/components/schemas/Group'
            - type: object
              properties:
                number_of_groups:
                  type:
                  - integer
                  - 'null'
                  format: int64
                  description: |-
                    The number of groups owned directly by this group

                    This information is only present when requested.
                  minimum: 0
                number_of_sboms:
                  type:
                  - integer
                  - 'null'
                  format: int64
                  description: |-
                    The number of SBOMs directly assigned to this group

                    This information is only present when requested.
                  minimum: 0
                parents:
                  type:
                  - array
                  - 'null'
                  items:
                    type: string
                  description: |-
                    The path, of IDs, from the root to this group

                    This information is only present when requested.
            description: Detailed group information, extends [`Group`]
        total:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
...